Made all functions nothrow
zmq_msg_init_data() takes a callback, which must also be nothrow if the
function itself is nothrow.  Alternatively, zmq_msg_init_data() *and*
the callback could be allowed to throw, but since the function is most
likely not written with this in mind, making everyhing nothrow seems
like the safest choice.